Animatable npm package for simple entrance animations

+++ components/FavoriteComponent.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
 import { baseUrl } from '../shared/baseUrl';
 import Swipeout from 'react-native-swipeout';
+import * as Animatable from 'react-native-animatable';
 
 import { deleteFavorite } from '../redux/ActionCreator';
 
@@ -36,19 +37,38 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ]
 
             return (
                 <Swipeout right={rightButton} autoClose={true}>
-                    <ListItem
-                        key={index}
-                        title={item.name}
-                        subtitle={item.description}
-                        onPress={() => navigate('DishDetail', { dishId: item.id })}
-                        leftAvatar={{ source: { uri: baseUrl + item.image } }}
-                    />
+                    <Animatable.View animation='fadeInRightBig' duration={2000}>
+                        <ListItem
+                            key={index}
+                            title={item.name}
+                            subtitle={item.description}
+                            onPress={() => navigate('DishDetail', { dishId: item.id })}
+                            leftAvatar={{ source: { uri: baseUrl + item.image } }}
+                        />
+                    </Animatable.View>
                 </Swipeout>
             );
         };
